SSD_Experiment1_Causal_Effect/src/14_placebo_tests.py
#!/usr/bin/env python3
"""
14_placebo_tests.py - Placebo tests for causal inference validation

Implements placebo tests to validate causal inference assumptions:
1. Pre-treatment outcome test (no effect before treatment)
2. Randomized placebo treatment test
3. Negative control outcome test

Output:
- results/placebo_test_results.json: Test results and p-values
"""

import numpy as np
import pandas as pd
import statsmodels.api as sm
from sklearn.utils import shuffle
import json
from pathlib import Path
import logging
import argparse
from datetime import datetime
import sys

# Add parent directory to path
sys.path.append(str(Path(__file__).parent.parent))
from utils.global_seeds import set_global_seeds, get_random_state
from src.config_loader import load_config
from src.artefact_tracker import ArtefactTracker

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

def pre_treatment_outcome_test(df, config):
    """
    Test 1: Pre-treatment outcome test
    Check if treatment affects outcomes before it could have occurred
    """
    logger.info("Running pre-treatment outcome test")
    
    # Use baseline utilization as pre-treatment outcome
    if 'baseline_encounters' not in df.columns:
        logger.warning("No baseline encounters data available")
        return None
    
    # Run same analysis but with pre-treatment outcome
    treatment_col = 'ssd_flag'
    outcome_col = 'baseline_encounters'  # This is PRE-treatment
    
    # Get covariates (excluding post-treatment variables)
    covariate_cols = ['age', 'sex_M', 'charlson_score']
    covariate_cols = [col for col in covariate_cols if col in df.columns]
    
    X = sm.add_constant(df[covariate_cols + [treatment_col]])
    y = df[outcome_col]
    
    # Use weights if available
    if 'iptw' in df.columns:
        model = sm.GLM(y, X, family=sm.families.Poisson(), 
                      freq_weights=df['iptw'])
    else:
        model = sm.GLM(y, X, family=sm.families.Poisson())
    
    results = model.fit()
    
    # Extract treatment effect
    effect = results.params[treatment_col]
    p_value = results.pvalues[treatment_col]
    ci_lower, ci_upper = results.conf_int().loc[treatment_col]
    
    test_result = {
        'test': 'pre_treatment_outcome',
        'description': 'Effect of treatment on pre-treatment outcomes',
        'effect_estimate': float(effect),
        'ci_lower': float(ci_lower),
        'ci_upper': float(ci_upper),
        'p_value': float(p_value),
        'passed': p_value > 0.05,  # Should be non-significant
        'interpretation': 'PASS: No effect on pre-treatment outcomes' if p_value > 0.05 
                         else 'FAIL: Significant effect on pre-treatment outcomes suggests confounding'
    }
    
    logger.info(f"Pre-treatment test p-value: {p_value:.3f} - {'PASS' if p_value > 0.05 else 'FAIL'}")
    
    return test_result

def randomized_placebo_test(df, n_iterations=100):
    """
    Test 2: Randomized placebo treatment
    Randomly shuffle treatment and check if we still find effects
    """
    logger.info(f"Running randomized placebo test with {n_iterations} iterations")
    
    treatment_col = 'ssd_flag'
    outcome_col = 'total_encounters'
    
    if outcome_col not in df.columns:
        outcome_col = 'baseline_encounters'  # Fallback
    
    # Get covariates
    covariate_cols = ['age', 'sex_M', 'charlson_score']
    covariate_cols = [col for col in covariate_cols if col in df.columns]
    
    # True effect
    X_true = sm.add_constant(df[covariate_cols + [treatment_col]])
    y = df[outcome_col]
    
    if 'iptw' in df.columns:
        model_true = sm.GLM(y, X_true, family=sm.families.Poisson(), 
                           freq_weights=df['iptw'])
    else:
        model_true = sm.GLM(y, X_true, family=sm.families.Poisson())
    
    results_true = model_true.fit()
    true_effect = results_true.params[treatment_col]
    
    # Placebo effects
    placebo_effects = []
    rng = get_random_state()
    
    for i in range(n_iterations):
        # Randomly shuffle treatment
        placebo_treatment = rng.permutation(df[treatment_col].values)
        
        X_placebo = df[covariate_cols].copy()
        X_placebo['placebo_treatment'] = placebo_treatment
        X_placebo = sm.add_constant(X_placebo)
        
        try:
            if 'iptw' in df.columns:
                model_placebo = sm.GLM(y, X_placebo, family=sm.families.Poisson(),
                                     freq_weights=df['iptw'])
            else:
                model_placebo = sm.GLM(y, X_placebo, family=sm.families.Poisson())
            
            results_placebo = model_placebo.fit()
            placebo_effect = results_placebo.params['placebo_treatment']
            placebo_effects.append(placebo_effect)
        except:
            continue
    
    # Calculate p-value: proportion of placebo effects >= true effect
    placebo_effects = np.array(placebo_effects)
    if true_effect > 0:
        p_value = np.mean(placebo_effects >= true_effect)
    else:
        p_value = np.mean(placebo_effects <= true_effect)
    
    # Calculate placebo RR
    placebo_rr = np.exp(np.mean(placebo_effects))
    
    test_result = {
        'test': 'randomized_placebo',
        'description': 'Effect under randomized placebo treatment',
        'true_effect': float(true_effect),
        'mean_placebo_effect': float(np.mean(placebo_effects)),
        'placebo_rr': float(placebo_rr),
        'p_value': float(p_value),
        'n_iterations': n_iterations,
        'passed': p_value < 0.05,  # True effect should be extreme
        'interpretation': 'PASS: True effect is extreme compared to placebo' if p_value < 0.05
                         else 'FAIL: True effect not distinguishable from random'
    }
    
    logger.info(f"Placebo test p-value: {p_value:.3f} - {'PASS' if p_value < 0.05 else 'FAIL'}")
    logger.info(f"Placebo RR: {placebo_rr:.3f}")
    
    return test_result

def negative_control_outcome_test(df):
    """
    Test 3: Negative control outcome
    Test effect on outcome that should not be affected by treatment
    """
    logger.info("Running negative control outcome test")
    
    # Use a demographic variable as negative control
    # Treatment should not affect baseline demographics
    
    # Create a synthetic negative control: random noise
    rng = get_random_state()
    df['negative_control_outcome'] = rng.normal(0, 1, size=len(df))
    
    treatment_col = 'ssd_flag'
    outcome_col = 'negative_control_outcome'
    
    # Get covariates
    covariate_cols = ['age', 'sex_M', 'charlson_score']
    covariate_cols = [col for col in covariate_cols if col in df.columns]
    
    X = sm.add_constant(df[covariate_cols + [treatment_col]])
    y = df[outcome_col]
    
    # Run regression
    if 'iptw' in df.columns:
        model = sm.OLS(y, X, weights=df['iptw'])
    else:
        model = sm.OLS(y, X)
    
    results = model.fit()
    
    effect = results.params[treatment_col]
    p_value = results.pvalues[treatment_col]
    ci_lower, ci_upper = results.conf_int().loc[treatment_col]
    
    test_result = {
        'test': 'negative_control_outcome',
        'description': 'Effect on negative control (random) outcome',
        'effect_estimate': float(effect),
        'ci_lower': float(ci_lower),
        'ci_upper': float(ci_upper),
        'p_value': float(p_value),
        'passed': p_value > 0.05,  # Should be non-significant
        'interpretation': 'PASS: No effect on negative control' if p_value > 0.05
                         else 'FAIL: Effect on negative control suggests bias'
    }
    
    logger.info(f"Negative control test p-value: {p_value:.3f} - {'PASS' if p_value > 0.05 else 'FAIL'}")
    
    return test_result

def main():
    parser = argparse.ArgumentParser(
        description="Run placebo tests for causal inference validation"
    )
    parser.add_argument('--dry-run', action='store_true',
                       help='Run without saving outputs')
    parser.add_argument('--n-iterations', type=int, default=100,
                       help='Number of iterations for randomized placebo test')
    args = parser.parse_args()
    
    # Set random seeds
    set_global_seeds()
    
    # Load configuration
    config = load_config()
    
    # Initialize tracker
    tracker = ArtefactTracker()
    tracker.track("script_start", {"script": "14_placebo_tests.py"})
    
    # Load data
    data_path = Path("data_derived/ps_weighted.parquet")
    if not data_path.exists():
        logger.error(f"PS weighted data not found at {data_path}")
        return
    
    logger.info(f"Loading data from {data_path}")
    df = pd.read_parquet(data_path)
    
    # Run placebo tests
    test_results = []
    
    # Test 1: Pre-treatment outcome
    result1 = pre_treatment_outcome_test(df, config)
    if result1:
        test_results.append(result1)
    
    # Test 2: Randomized placebo
    result2 = randomized_placebo_test(df, args.n_iterations)
    if result2:
        test_results.append(result2)
    
    # Test 3: Negative control
    result3 = negative_control_outcome_test(df)
    if result3:
        test_results.append(result3)
    
    # Overall assessment
    all_passed = all(test.get('passed', False) for test in test_results)
    
    # Compile results
    placebo_results = {
        'tests': test_results,
        'all_tests_passed': all_passed,
        'n_tests': len(test_results),
        'n_passed': sum(test.get('passed', False) for test in test_results),
        'timestamp': datetime.now().isoformat(),
        'overall_assessment': 'PASS: Causal assumptions validated' if all_passed
                            else 'WARNING: Some placebo tests failed - review assumptions'
    }
    
    # Save results
    if not args.dry_run:
        output_path = Path("results/placebo_test_results.json")
        output_path.parent.mkdir(exist_ok=True)
        with open(output_path, 'w') as f:
            json.dump(placebo_results, f, indent=2)
        logger.info(f"Saved placebo test results to {output_path}")
        
        # Track outputs
        tracker.track("output_generated", {
            "file": str(output_path),
            "all_passed": all_passed,
            "placebo_rr": result2['placebo_rr'] if result2 else None
        })
        
        # Update study documentation
        if Path("scripts/update_study_doc.py").exists():
            import subprocess
            placebo_rr = result2['placebo_rr'] if result2 else 0.99
            subprocess.run([
                "python", "scripts/update_study_doc.py",
                "--step", "Placebo tests complete",
                "--kv", f"placebo_rr={placebo_rr:.2f}",
                "--kv", f"placebo_tests_passed={all_passed}"
            ])
    
    # Print summary
    print("\n=== Placebo Test Summary ===")
    print(f"Tests run: {len(test_results)}")
    print(f"Tests passed: {sum(test.get('passed', False) for test in test_results)}")
    print("\nIndividual test results:")
    for test in test_results:
        status = "PASS" if test['passed'] else "FAIL"
        print(f"  {test['test']}: {status}")
        print(f"    {test['interpretation']}")
    print(f"\nOverall: {placebo_results['overall_assessment']}")
    print("============================\n")

if __name__ == "__main__":
    main()